LayoutTests/platform/chromium-win/fast/workers/worker-close-expected.txt

Test WorkerContext.close functionality. Should print a series of PASS messages, followed with DONE.

PASS: typeof close: function
PASS: received message before close
PASS: messages sent after close() are ignored
PASS: Error arrived after close: Uncaught ReferenceError: nonExistentFunction is not defined
PASS: close() did not dispatch pending events
DONE
